Initial code for 3D boundary normals
