+++ deep_conmech/simulator/setting/matrices_3d.py
